# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "一个简单的机器管理系统",
        "title": "rabbit",
        "contact": {},
        "license": {}
    },
    "paths": {
...
        "/api/v1/template/create": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "创建新模板",
                "parameters": [
                    {
                        "description": "创建新模板",
                        "name": "APIPostTemplateUpdateInputs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app.APIPostTemplateUpdateInputs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.APIPostTemplateUpdateInputs"
                        }
                    },
                    "400": {
                        "description": "bad arguments"
                    },
                    "417": {
                        "description": "internal error"
                    }
                }
            }
        },
        "/api/v1/template/design": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "模板设计",
                "parameters": [
                    {
                        "description": "模板设计",
                        "name": "APIPostTemplateDesignInputs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app.APIPostTemplateDesignInputs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Template"
                        }
                    },
                    "400": {
                        "description": "bad parameters"
                    },
                    "417": {
                        "description": "internal error"
                    }
                }
            }
        },
        "/api/v1/template/info": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "根据机器ID获取模板详细信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "根据机器ID获取模板详细信息",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Template"
                        }
                    },
                    "400": {
                        "description": "bad arguments"
                    },
                    "417": {
                        "description": "internal error"
                    }
                }
            }
        },
        "/api/v1/template/list": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "展现模板列表接口",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "orderBy",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "remark",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.APIGetTemplateListOutputs"
                        }
                    },
                    "400": {
                        "description": "bad arguments"
                    },
                    "417": {
                        "description": "internal error"
                    }
                }
            }
        },
        "/api/v1/template/update": {
            "put": {
                "produces": [
                    "application/json"
                ],
                "summary": "更新模板信息",
                "parameters": [
                    {
                        "description": "更新模板信息",
                        "name": "APIPostTemplateUpdateInputs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app.APIPostTemplateUpdateInputs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Template"
                        }
                    },
                    "400": {
                        "description": "bad parameters"
                    },
                    "417": {
                        "description": "internal error"
                    }
                }
            }
        },
        "/api/v1/tree": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "获取host group信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "根据ID获取host group树状信息",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.APIGetHostGroupTreeOutputs"
                        }
                    },
                    "400": {
                        "description": "bad arguments"
                    },
                    "417": {
                        "description": "internal error"
                    }
                }
            }
        },
...
        "/api/v2/tree": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "根据tags获取tags下所有的机器",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "根据tags获取tags下所有的机器",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.APIGetHostGroupTreeOutputs"
                        }
                    },
                    "400": {
                        "description": "bad arguments"
                    },
                    "417": {
                        "description": "internal error"
                    }
                }
            }
        },
        "/api/v3/tree": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "V3版根据tags获取tags下所有的机器",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "根据tags获取tags下所有的机器",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.APIGetHostGroupTreeOutputs"
                        }
                    },
                    "400": {
                        "description": "bad arguments"
                    },
                    "417": {
                        "description": "internal error"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "app.APIGetHostGroupTreeOutputs": {
            "type": "object",
            "properties": {
                "hostGroups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/node.HostGroup"
                    }
                }
            }
        },
...
        "app.APIGetTemplateListOutputs": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/app.Template"
                    }
                },
                "totalCount": {
                    "type": "integer"
                }
            }
        },
...
        "app.APIPostTemplateDesignInputs": {
            "type": "object",
            "properties": {
                "edges": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/app.G6Edge"
                    }
                },
                "groups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/app.G6Group"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "nodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/app.G6Node"
                    }
                }
            }
        },
        "app.APIPostTemplateUpdateInputs": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                }
            }
        },
        "app.G6Edge": {
            "type": "object",
            "properties": {
                "end": {
                    "type": "object",
                    "$ref": "#/definitions/app.G6Point"
                },
                "endPoint": {
                    "type": "object",
                    "$ref": "#/definitions/app.G6Point"
                },
                "id": {
                    "type": "string"
                },
                "shape": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "sourceId": {
                    "description": "edge.sourceID 指向 node.id",
                    "type": "integer"
                },
                "start": {
                    "type": "object",
                    "$ref": "#/definitions/app.G6Point"
                },
                "startPoint": {
                    "type": "object",
                    "$ref": "#/definitions/app.G6Point"
                },
                "target": {
                    "type": "string"
                },
                "targetId": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "app.G6Group": {
            "type": "object"
        },
        "app.G6Node": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "inPoints": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "number"
                        }
                    }
                },
                "label": {
                    "type": "string"
                },
                "name": {
                    "description": "TagID      int64       `json:\"tagID\" form:\"tagID\"`",
                    "type": "string"
                },
                "offsetX": {
                    "type": "integer"
                },
                "offsetY": {
                    "type": "integer"
                },
                "outPoints": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "number"
                        }
                    }
                },
                "shape": {
                    "type": "string"
                },
                "size": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "stateImage": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "x": {
                    "type": "integer"
                },
                "y": {
                    "type": "integer"
                }
            }
        },
        "app.G6Point": {
            "type": "object",
            "properties": {
                "x": {
                    "type": "number"
                },
                "y": {
                    "type": "number"
                }
            }
        },
...
        "app.Template": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "createAt": {
                    "type": "object",
                    "$ref": "#/definitions/gtime.GTime"
                },
                "creator": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updateAt": {
                    "type": "object",
                    "$ref": "#/definitions/gtime.GTime"
                }
            }
        },
...
        "gtime.GTime": {
            "type": "object"
        },
...
        "node.HostGroup": {
            "type": "object",
            "properties": {
                "caasServiceId": {
                    "type": "integer"
                },
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/node.HostGroup"
                    }
                },
                "childrenHostCount": {
                    "type": "integer"
                },
                "childrenPodCount": {
                    "type": "integer"
                },
                "createUser": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isWarning": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "parentID": {
                    "type": "integer"
                },
                "parentName": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "pathArray": {
                    "type": "string"
                },
                "relatedHostCount": {
                    "type": "integer"
                },
                "relatedPodCount": {
                    "type": "integer"
                },
                "subGroupCount": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
    }
}